+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [?] - ?
+
+### Added
+- Few default fog settings.
+
+### Changed
+- Few default fog settings.
+- Cleared the project a little. Removed some unneeded code.
+
 ## [1.1.1] - 2021-03-22
 
 ### Added